ath79: add support for AVM Fritzbox 3490, 5490 and 7490 WASP SoC
This adds support for the Fritzbox 3490, 5490 and 7490 WASP SoC
(Wireless Assistant Support Processor, sometimes referred to as
Scorpion).
This is a memory only target on the mentioned Lantiq devices, that
has no access to leds, NAND, GPIO, USB3 on the Lantiq host.
The two wifi adapters with 2.4GHz and 5 GHz are connected to this
SoC and they are not accessible from the Lantiq host. The ethernet
port is connected to the switch on the Lantiq host and allows
network communication. The Lantiq can load a network boot loader
to this SoC using gpio.
The WASP SoC runs fully from RAM and is booted by using a remoteproc
kernel module and runs its own OpenWrt instance.

Specifications:
 - SoC: QCA9558 on Lantiq host
 - RAM: 128 MB
 - Wireless:
   · Qualcomm-QCA9558 w/ 3×3 MIMO for 2.4GHz 802.11b/g/n
   · Qualcomm-QCA9880 w/ 3×3 MIMO for 5GHz 802.11a/ac
 - AG71xx ethernet

Installation:
For booting WASP, build the ATH79 target x7490 WASP image, copy it
to /lib/firmware on the Lantiq host image and use the network boot
firmware from the AVM firmware together with the avm_wasp kernel
module.

Signed-off-by: Daniel Kestrel <kestrel1974@t-online.de>
(cherry picked from commit 40000f06371f399ea55e9dbc4aa065e54be16c10)